+++ internal/models/Order.go
@@ -8,7 +8,7 @@ type Order struct {
 	ID         int       `json:"id"`
 	CustomerID int       `json:"customer_id"`
 	Status     string    `json:"status"`
-	Accrual    int       `json:"accrual"`
+	Accrual    float64   `json:"accrual"`
 	UploadedAt time.Time `json:"uploaded_at"`
 }
 
@@ -16,7 +16,7 @@ type Order struct {
 type ResponseGetOwnerOrders struct {
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    int       `json:"accrual"`
+	Accrual    float64   `json:"accrual"`
 	UploadedAt time.Time `json:"uploadedAt"`
 }
